+++ hyperkitty/tests/test_tasks.py
@@ -39,6 +39,12 @@ class TaskTestCase(TestCase):
         except Thread.DoesNotExist:
             self.fail("No protection when the thread is deleted")
 
+    def test_rebuild_thread_cache_votes_no_thread(self):
+        try:
+            tasks.rebuild_thread_cache_votes(42)
+        except Thread.DoesNotExist:
+            self.fail("No protection when the thread is deleted")
+
     def test_compute_thread_positions_no_thread(self):
         try:
             tasks.compute_thread_positions(42)